refactor parse_keywords.rs
