+++ crates/nu-parser/src/lib.rs
@@ -5,16 +5,21 @@ mod flatten;
 mod known_external;
 mod lex;
 mod lite_parser;
+mod parse_alias;
+mod parse_bindings;
 mod parse_calls;
 mod parse_captures_compile;
+mod parse_def;
 mod parse_expressions;
 mod parse_helpers;
 mod parse_keywords;
 mod parse_literals;
+mod parse_module;
 mod parse_patterns;
 mod parse_pipelines;
 mod parse_shape_specs;
 mod parse_signatures;
+mod parse_source;
 mod parser;
 mod type_check;
